Added timer + physics + 1/2 collisions(commit for analysis)

+++ src/Game/GameObjects/BetonWall.h
@@ -35,7 +35,7 @@ public:
     };
     BetonWall(const EBetonWallType& eBetonWallType, const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
      void Render() const override; // убрал виртуал........................................................
-     void update(const uint64_t& delta) override;
+     void update(const double& delta) override;
 
 private:
     void RenderBlock(const EBlockLocation& eBlockLocation) const;
