+++ src/Game/GameObjects/Water.cpp
@@ -30,7 +30,7 @@ void Water::Render() const
 		RenderBlock(EBlockLocation::BottomRight);
 }
 
-void Water::update(const uint64_t& delta)
+void Water::update(const double& delta)
 {
 	m_spriteAnimator.update(delta);
 }
