+++ src/Game/GameObjects/Water.h
@@ -19,7 +19,7 @@ public:
     };
     Water(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
      void Render() const override;  
-     void update(const uint64_t& delta) override;
+     void update(const double& delta) override;
     
 
 private:
